Резервное копирование

+++ src/database/DataBaseManager.cpp
@@ -6,6 +6,7 @@
 #include <QTextCodec>
 #include "SimpleIni.h"
 #include "ServerSettingUtils.h"
+#include <QMessageBox>
 
 qint64 DataBaseManager::idStartLastDuty = -1;
 
@@ -283,9 +284,30 @@ QSqlDatabase& DataBaseManager::m_db()
     return db;
 }
 
-void DataBaseManager::makeBackup()
+bool DataBaseManager::makeBackup(QString backup)
 {
 
+    //Проверить указана ли папка для бэкапа
+    //Проверить есть ли физически эта папка
+
+    //Если нет вернуть false
+    if(!QDir(DataBaseManager::BackupPath).exists()){
+
+        QMessageBox msgBox;
+        msgBox.setText("Директория для бэкапа отсутствует");
+        msgBox.exec();
+
+        return false;
+    }
+
+    if(!QDir(DataBaseManager::BackupPath).exists()){
+
+        QMessageBox msgBox;
+        msgBox.setText("Директория для бэкапа отсутствует");
+        msgBox.exec();
+
+        return false;
+    }
 
     qDebug()<<"BackupPath  : "<<DataBaseManager::BackupPath;
     qDebug()<<"DatabaseName: "<<DataBaseManager::DatabaseName;
@@ -294,7 +316,7 @@ void DataBaseManager::makeBackup()
     qDebug()<<"Port        : "<<DataBaseManager::Port;
     qDebug()<<"UserName    : "<<DataBaseManager::UserName;
 
-
+    int res;
 
  //   QString cmd="/usr/bin/pg_dump --host 127.0.0.1 --port 5432 --username postgres --no-password --format custom --blobs --verbose --file backup_2 rif_db0 ";
 
@@ -302,30 +324,86 @@ void DataBaseManager::makeBackup()
              " --username "+ DataBaseManager::UserName+
              " --no-password --format custom --blobs --verbose --file backup "+ DataBaseManager::DatabaseName;
     qDebug()<<cmd;
-   QProcess::execute(cmd);
+    res=QProcess::execute(cmd);
+    qDebug()<<"res: "<<res;
 
+    if(res!=0){
 
-    cmd="zip --password MY_SECRET user.zip rifx.ini backup";
+        QMessageBox msgBox;
+        msgBox.setText("Резервное копирование не выполнено!!");
+        msgBox.exec();
+        return false;
+    }
+
+    cmd="zip --password MY_SECRET "+backup+".zip rifx.ini backup";
     qDebug()<<cmd;
-    QProcess::execute(cmd);
+    res=QProcess::execute(cmd);
+    qDebug()<<"res: "<<res;
 
-    cmd="cp user.zip "+DataBaseManager::BackupPath;
+    if(res!=0){
+
+        QMessageBox msgBox;
+        msgBox.setText("Резервное копирование не выполнено!!");
+        msgBox.exec();
+        return false;
+    }
+
+    cmd="cp "+backup+".zip "+DataBaseManager::BackupPath;
     qDebug()<<cmd;
-    QProcess::execute(cmd);
+    res=QProcess::execute(cmd);
+    qDebug()<<"res: "<<res;
+
+    if(res!=0){
+
+        QMessageBox msgBox;
+        msgBox.setText("Резервное копирование не выполнено!!");
+        msgBox.exec();
+        return false;
+    }
 
     cmd="rm backup";
     qDebug()<<cmd;
-    QProcess::execute(cmd);
+    res=QProcess::execute(cmd);
+    qDebug()<<"res: "<<res;
 
-    cmd="rm user.zip";
+    if(res!=0){
+
+        QMessageBox msgBox;
+        msgBox.setText("Резервное копирование не выполнено!!");
+        msgBox.exec();
+        return false;
+    }
+
+    cmd="rm "+backup+".zip";
     qDebug()<<cmd;
-    QProcess::execute(cmd);
+    res=QProcess::execute(cmd);
+    qDebug()<<"res: "<<res;
 
+    if(res!=0){
+
+        QMessageBox msgBox;
+        msgBox.setText("Резервное копирование не выполнено!!");
+        msgBox.exec();
+        return false;
+    }
+
+    QMessageBox msgBox;
+    msgBox.setText("Резервное копирование завершено");
+    msgBox.exec();
+    return true;
 
 }
 
-void DataBaseManager::restoreBackup()
+bool DataBaseManager::restoreBackup(QString backup)
 {
+
+    //Проверить указана ли папка для бэкапа
+    //Проверить есть ли физически эта папка
+
+    //Проверить есть ли в ней этот архив
+
+    //Если нет вернуть false
+
     qDebug()<<"BackupPath  : "<<DataBaseManager::BackupPath;
     qDebug()<<"DatabaseName: "<<DataBaseManager::DatabaseName;
     qDebug()<<"HostName    : "<<DataBaseManager::HostName;
@@ -334,14 +412,20 @@ void DataBaseManager::restoreBackup()
     qDebug()<<"UserName    : "<<DataBaseManager::UserName;
 
 
-
+int res;
  //   QString cmd="/usr/bin/pg_restore --host 127.0.0.1 --port 5432 --username "postgres" --dbname "rif_db3" --role "postgres" --no-password  --verbose "/home/administrator/Backup/backup"";
 
-    QString cmd="unzip -P MY_SECRET "+DataBaseManager::BackupPath+"/user.zip -d "+DataBaseManager::BackupPath;
+    QString cmd="unzip -P MY_SECRET "+DataBaseManager::BackupPath+"/"+backup+".zip -d "+DataBaseManager::BackupPath;
     qDebug()<<cmd;
-    QProcess::execute(cmd);
+    res=QProcess::execute(cmd);
 
+    if(res!=0){
 
+        QMessageBox msgBox;
+        msgBox.setText("Восстановление не выполнено!!");
+        msgBox.exec();
+        return false;
+    }
 
 
 cmd="/usr/bin/pg_restore --host 127.0.0.1 --port "+DataBaseManager::Port+
@@ -350,79 +434,58 @@ cmd="/usr/bin/pg_restore --host 127.0.0.1 --port "+DataBaseManager::Port+
              " --role "+DataBaseManager::UserName+
              "  --no-password  --verbose "+DataBaseManager::BackupPath+"/backup";
 
+
   qDebug()<<cmd;
-  QProcess::execute(cmd);
+  res=QProcess::execute(cmd);
+  qDebug()<<"res: "<<res;
+  if((res!=0)&&(res!=1)){
+
+      cmd="rm "+DataBaseManager::BackupPath+"/backup";
+      res=QProcess::execute(cmd);
+
+      cmd="rm "+DataBaseManager::BackupPath+"/rifx.ini";
+      res=QProcess::execute(cmd);
+
+      QMessageBox msgBox;
+      msgBox.setText("Восстановление не выполнено!!");
+      msgBox.exec();
+      return false;
+  }
 
   cmd="cp "+DataBaseManager::BackupPath+"/rifx.ini rifx.ini ";
   qDebug()<<cmd;
-  QProcess::execute(cmd);
+  res=QProcess::execute(cmd);
+  qDebug()<<"res: "<<res;
+  if(res!=0){
+
+      cmd="rm "+DataBaseManager::BackupPath+"/backup";
+      res=QProcess::execute(cmd);
+
+      cmd="rm "+DataBaseManager::BackupPath+"/rifx.ini";
+      res=QProcess::execute(cmd);
+
+      QMessageBox msgBox;
+      msgBox.setText("Восстановление не выполнено!!");
+      msgBox.exec();
+      return false;
+  }
 
   cmd="rm "+DataBaseManager::BackupPath+"/backup";
-  qDebug()<<cmd;
-  QProcess::execute(cmd);
+  res=QProcess::execute(cmd);
 
   cmd="rm "+DataBaseManager::BackupPath+"/rifx.ini";
-  qDebug()<<cmd;
-  QProcess::execute(cmd);
+  res=QProcess::execute(cmd);
 
+
+  QMessageBox msgBox;
+  msgBox.setText("Восстановление завершено.Перезапустите сервер.");
+  msgBox.exec();
+  return true;
 }
 
-void DataBaseManager::makeAutoBackup()
-{
-    QString cmd="/usr/bin/pg_dump --host 127.0.0.1 --port "+DataBaseManager::Port+
-                " --username "+ DataBaseManager::UserName+
-                " --no-password --format custom --blobs --verbose --file backup "+ DataBaseManager::DatabaseName;
-       qDebug()<<cmd;
-      QProcess::execute(cmd);
-
-
-       cmd="zip --password MY_SECRET auto.zip rifx.ini backup";
-       qDebug()<<cmd;
-       QProcess::execute(cmd);
-
-       cmd="cp auto.zip "+DataBaseManager::BackupPath;
-       qDebug()<<cmd;
-       QProcess::execute(cmd);
-
-       cmd="rm backup";
-       qDebug()<<cmd;
-       QProcess::execute(cmd);
-
-       cmd="rm auto.zip";
-       qDebug()<<cmd;
-       QProcess::execute(cmd);
-}
-
-void DataBaseManager::restoreAutoBackup()
-{
-    QString cmd="unzip -P MY_SECRET "+DataBaseManager::BackupPath+"/auto.zip -d "+DataBaseManager::BackupPath;
-    qDebug()<<cmd;
-    QProcess::execute(cmd);
 
 
 
-
-cmd="/usr/bin/pg_restore --host 127.0.0.1 --port "+DataBaseManager::Port+
-             " --username "+ DataBaseManager::UserName+
-             " --dbname "+DataBaseManager::DatabaseName+
-             " --role "+DataBaseManager::UserName+
-             "  --no-password  --verbose "+DataBaseManager::BackupPath+"/backup";
-
-  qDebug()<<cmd;
-  QProcess::execute(cmd);
-
-  cmd="cp "+DataBaseManager::BackupPath+"/rifx.ini rifx.ini ";
-  qDebug()<<cmd;
-  QProcess::execute(cmd);
-
-  cmd="rm "+DataBaseManager::BackupPath+"/backup";
-  qDebug()<<cmd;
-  QProcess::execute(cmd);
-
-  cmd="rm "+DataBaseManager::BackupPath+"/rifx.ini";
-  qDebug()<<cmd;
-  QProcess::execute(cmd);
-}
 
 DataBaseManager::~DataBaseManager() noexcept
 {
